Rename inst_monitor.% to object_monitor.%
