+++ ruby/bop/object_monitor.h
