+++ ruby/common.mk
@@ -102,7 +102,7 @@ COMMONOBJS    = array.$(OBJEXT) \
 		thread.$(OBJEXT) \
 		cont.$(OBJEXT) \
 		bop/ppr.$(OBJEXT) \
-		bop/inst_monitor.$(OBJEXT) \
+		bop/object_monitor.$(OBJEXT) \
 		$(DTRACE_OBJ) \
 		$(BUILTIN_ENCOBJS) \
 		$(BUILTIN_TRANSOBJS) \
@@ -2398,10 +2398,10 @@ vm_trace.$(OBJEXT): {$(VPATH)}vm_trace.c
 # AUTOGENERATED DEPENDENCIES END
 
 #MANUALLY ADDED BOP DEPENDENCIES
-bop/inst_monitor.$(OBJEXT): $(top_srcdir)/bop/inst_monitor.h
-bop/inst_monitor.$(OBJEXT): $(top_srcdir)/bop/ppr.h
+bop/object_monitor.$(OBJEXT): $(top_srcdir)/bop/object_monitor.h
+bop/object_monitor.$(OBJEXT): $(top_srcdir)/bop/ppr.h
 
-bop/ppr.$(OBJEXT): $(top_srcdir)/bop/inst_monitor.h
+bop/ppr.$(OBJEXT): $(top_srcdir)/bop/object_monitor.h
 bop/ppr.$(OBJEXT): $(top_srcdir)/bop/ppr.h
 
 gc.$(OBJEXT): $(top_srcdir)/bop/ppr.h
